+++ src/components/loader/loader.module.scss
@@ -19,5 +19,4 @@
   display: grid;
   justify-content: center;
   align-items: center;
-  height: 350px;
 }
